python-agent/requirements.txt
chainer
ws4py
cherrypy
msgpack-python
pillow
argparse